fix: repaired buttons, remove events from game struct, verifying if update function exists before calling it in update_scene, add docs

+++ src/events/analyse_events.c
@@ -14,7 +14,7 @@
 
 static void trigger_associated_ev(sfEvent *evt, game_t *game)
 {
-    for (int i = 0; evts_types[i].type != sfEvtCount; i++)
+    for (size_t i = 0; evts_types[i].type != sfEvtCount; i++)
         if (evts_types[i].fnct && evts_types[i].type == evt->type)
             evts_types[i].fnct(evt, game);
 }
